Add Authenticator trait
