+++ src/main/scala/auth/JwtAuthenticator.scala
@@ -14,11 +14,8 @@ import scala.util.Try
 
 case class JwtAuthenticator[T] (
   identityClient: IdentityClient,
-  private val baseEndpoint: PublicEndpoint[Unit, Unit, Unit, Any] = endpoint,
   private val jwtProcessor: Jwt => IO[Either[String, T]] = jwt => IO.pure(Right(jwt)),
-) {
-  def withBaseEndpoint(endpoint: PublicEndpoint[Unit, Unit, Unit, Any]): JwtAuthenticator[T] =
-    copy(baseEndpoint = endpoint)
+) extends Authenticator[T] {
 
   def requireAppJwt: JwtAuthenticator[AppJwt] = require(_.asApp)
   def requireAnyUserJwt: JwtAuthenticator[AnyUserJwt] = require(_.asAnyUser)
@@ -36,22 +33,6 @@ case class JwtAuthenticator[T] (
   def flatMapJwtF[R](f: T => IO[Either[String, R]]): JwtAuthenticator[R] =
     copy(jwtProcessor = jwtProcessor.andThen(e => EitherT(e).flatMap(t => EitherT(f(t))).value))
 
-  def secureEndpoint: PartialServerEndpoint[Option[String], T, Unit, AuthException, Unit, Any, IO] =
-    baseEndpoint
-      .securityIn(header[Option[String]]("Authorization"))
-      .errorOut(statusCode(StatusCode.Unauthorized).and(jsonBody[String].mapTo[AuthException]))
-      .serverSecurityLogic(authenticateHeader(_).map(_.left.map(AuthException(_))))
-
-  def authenticateHeader(headerOption: Option[String]): IO[Either[String, T]] =
-    headerOption match
-      case Some(header) => authenticateHeader(header)
-      case None => IO.pure(Left("Missing Authorization header"))
-
-  def authenticateHeader(header: String): IO[Either[String, T]] =
-    EitherT.fromEither[IO](extractTokenFromHeader(header))
-      .flatMap(token => EitherT(authenticateToken(token)))
-      .value
-
   def authenticateToken(raw: String): IO[Either[String, T]] =
     (for {
       token <- EitherT.fromEither[IO](parseToken(raw))
@@ -60,12 +41,6 @@ case class JwtAuthenticator[T] (
       jwt <- EitherT.pure[IO, String](Jwt.fromJavaToken(token))
       mapped <- EitherT(jwtProcessor(jwt))
     } yield mapped).value
-
-  private def extractTokenFromHeader(header: String): Either[String, String] =
-    header match {
-      case s"Bearer $jwt" => Right(jwt)
-      case _              => Left("Invalid Authorization header")
-    }
 
   private def parseToken(raw: String): Either[String, Token] =
     Try(Token.parse(raw)).toEither.left.map(_.getMessage)
